+++ intg-appletv/setup_flow.py
@@ -35,18 +35,49 @@ class SetupSteps(IntEnum):
 
     INIT = 0
     CONFIGURATION_MODE = 1
-    DEVICE_CHOICE = 2
-    PAIRING_AIRPLAY = 3
-    PAIRING_COMPANION = 4
+    DISCOVER = 2
+    DEVICE_CHOICE = 3
+    PAIRING_AIRPLAY = 4
+    PAIRING_COMPANION = 5
 
 
 _setup_step = SetupSteps.INIT
+_cfg_add_device: bool = False
 _manual_address: bool = False
 _discovered_atvs: list[pyatv.interface.BaseConfig] = None
 _pairing_apple_tv: tv.AppleTv | None = None
+# TODO #12 externalize language texts
+# pylint: disable=line-too-long
+_user_input_discovery = RequestUserInput(
+    {"en": "Setup mode", "de": "Setup Modus"},
+    [
+        {
+            "id": "info",
+            "label": {
+                "en": "Discover or connect to Apple TV device",
+                "de": "Suche oder Verbinde auf Apple TV Gerät",
+                "fr": "Découvrir ou connexion à l'appareil Apple TV",
+            },
+            "field": {
+                "label": {
+                    "value": {
+                        "en": "Leave blank to use auto-discovery and click _Next_.",
+                        "de": "Leer lassen, um automatische Erkennung zu verwenden und auf _Weiter_ klicken.",
+                        "fr": "Laissez le champ vide pour utiliser la découverte automatique et cliquez sur _Suivant_.",  # noqa: E501
+                    }
+                }
+            },
+        },
+        {
+            "field": {"text": {"value": ""}},
+            "id": "address",
+            "label": {"en": "IP address", "de": "IP-Adresse", "fr": "Adresse IP"},
+        },
+    ],
+)
 
 
-async def driver_setup_handler(msg: SetupDriver) -> SetupAction:
+async def driver_setup_handler(msg: SetupDriver) -> SetupAction:  # pylint: disable=too-many-return-statements
     """
     Dispatch driver setup requests to corresponding handlers.
 
@@ -56,21 +87,26 @@ async def driver_setup_handler(msg: SetupDriver) -> SetupAction:
     :return: the setup action on how to continue
     """
     global _setup_step
+    global _cfg_add_device
     global _pairing_apple_tv
 
     if isinstance(msg, DriverSetupRequest):
         _setup_step = SetupSteps.INIT
-        return await handle_driver_setup(msg)
+        _cfg_add_device = False
+        return await _handle_driver_setup(msg)
+
     if isinstance(msg, UserDataResponse):
         _LOG.debug("%s", msg)
-        if _setup_step == SetupSteps.CONFIGURATION_MODE and "address" in msg.input_values:
-            return await handle_configuration_mode(msg)
+        if _setup_step == SetupSteps.CONFIGURATION_MODE and "action" in msg.input_values:
+            return await _handle_configuration_mode(msg)
+        if _setup_step == SetupSteps.DISCOVER and "address" in msg.input_values:
+            return await _handle_discovery(msg)
         if _setup_step == SetupSteps.DEVICE_CHOICE and "choice" in msg.input_values:
-            return await handle_device_choice(msg)
+            return await _handle_device_choice(msg)
         if _setup_step == SetupSteps.PAIRING_AIRPLAY and "pin_airplay" in msg.input_values:
-            return await handle_user_data_airplay_pin(msg)
+            return await _handle_user_data_airplay_pin(msg)
         if _setup_step == SetupSteps.PAIRING_COMPANION and "pin_companion" in msg.input_values:
-            return await handle_user_data_companion_pin(msg)
+            return await _handle_user_data_companion_pin(msg)
         _LOG.error("No or invalid user response was received: %s", msg)
     elif isinstance(msg, AbortDriverSetup):
         _LOG.info("Setup was aborted with code: %s", msg.error)
@@ -86,51 +122,139 @@ async def driver_setup_handler(msg: SetupDriver) -> SetupAction:
     return SetupError()
 
 
-async def handle_driver_setup(_msg: DriverSetupRequest) -> RequestUserInput | SetupError:
+async def _handle_driver_setup(msg: DriverSetupRequest) -> RequestUserInput | SetupError:
     """
     Start driver setup.
 
-    Initiated by Remote Two to set up the driver.
-    Ask user to enter ip-address for manual configuration, otherwise auto-discovery is used.
+    Initiated by Remote Two to set up the driver. The reconfigure flag determines the setup flow:
 
-    :param _msg: not used, we don't have any input fields in the first setup screen.
+    - Reconfigure is True: show the configured devices and ask user what action to perform (add, delete, reset).
+    - Reconfigure is False: clear the existing configuration and show device discovery screen.
+      Ask user to enter ip-address for manual configuration, otherwise auto-discovery is used.
+
+    :param msg: driver setup request data, only `reconfigure` flag is of interest.
     :return: the setup action on how to continue
     """
     global _setup_step
 
-    _LOG.debug("Starting driver setup")
-    _setup_step = SetupSteps.CONFIGURATION_MODE
-    # pylint: disable=line-too-long
-    return RequestUserInput(
-        {"en": "Setup mode", "de": "Setup Modus"},
-        [
+    reconfigure = msg.reconfigure
+    _LOG.debug("Starting driver setup, reconfigure=%s", reconfigure)
+
+    if reconfigure:
+        _setup_step = SetupSteps.CONFIGURATION_MODE
+
+        # get all configured devices for the user to choose from
+        dropdown_devices = []
+        for device in config.devices.all():
+            dropdown_devices.append({"id": device.identifier, "label": {"en": f"{device.name} ({device.identifier})"}})
+
+        # TODO #12 externalize language texts
+        # build user actions, based on available devices
+        dropdown_actions = [
             {
-                "id": "info",
+                "id": "add",
                 "label": {
-                    "en": "Discover or connect to Apple TV device",
-                    "de": "Suche oder Verbinde auf Apple TV Gerät",
-                    "fr": "Découvrir ou connexion à l'appareil Apple TV",
+                    "en": "Add a new device",
+                    "de": "Neues Gerät hinzufügen",
+                    "fr": "Ajouter un nouvel appareil",
                 },
-                "field": {
+            },
+        ]
+
+        # add remove & reset actions if there's at least one configured device
+        if dropdown_devices:
+            dropdown_actions.append(
+                {
+                    "id": "remove",
                     "label": {
-                        "value": {
-                            "en": "Leave blank to use auto-discovery and click _Next_.",
-                            "de": "Leer lassen, um automatische Erkennung zu verwenden und auf _Weiter_ klicken.",
-                            "fr": "Laissez le champ vide pour utiliser la découverte automatique et cliquez sur _Suivant_.",  # noqa: E501
-                        }
-                    }
+                        "en": "Delete selected device",
+                        "de": "Selektiertes Gerät löschen",
+                        "fr": "Supprimer l'appareil sélectionné",
+                    },
                 },
-            },
-            {
-                "field": {"text": {"value": ""}},
-                "id": "address",
-                "label": {"en": "IP address", "de": "IP-Adresse", "fr": "Adresse IP"},
-            },
-        ],
-    )
+            )
+            dropdown_actions.append(
+                {
+                    "id": "reset",
+                    "label": {
+                        "en": "Reset configuration and reconfigure",
+                        "de": "Konfiguration zurücksetzen und neu konfigurieren",
+                        "fr": "Réinitialiser la configuration et reconfigurer",
+                    },
+                },
+            )
+        else:
+            # dummy entry if no devices are available
+            dropdown_devices.append({"id": "", "label": {"en": "---"}})
+
+        return RequestUserInput(
+            {"en": "Configuration mode", "de": "Konfigurations-Modus"},
+            [
+                {
+                    "field": {"dropdown": {"value": dropdown_devices[0]["id"], "items": dropdown_devices}},
+                    "id": "choice",
+                    "label": {
+                        "en": "Configured devices",
+                        "de": "Konfigurierte Geräte",
+                        "fr": "Appareils configurés",
+                    },
+                },
+                {
+                    "field": {"dropdown": {"value": dropdown_actions[0]["id"], "items": dropdown_actions}},
+                    "id": "action",
+                    "label": {
+                        "en": "Action",
+                        "de": "Aktion",
+                        "fr": "Appareils configurés",
+                    },
+                },
+            ],
+        )
+
+    # Initial setup, make sure we have a clean configuration
+    config.devices.clear()  # triggers device instance removal
+    _setup_step = SetupSteps.DISCOVER
+    return _user_input_discovery
 
 
-async def handle_configuration_mode(msg: UserDataResponse) -> RequestUserInput | SetupError:
+async def _handle_configuration_mode(msg: UserDataResponse) -> RequestUserInput | SetupComplete | SetupError:
+    """
+    Process user data response from the configuration mode screen.
+
+    User input data:
+
+    - ``choice`` contains identifier of selected device
+    - ``action`` contains the selected action identifier
+
+    :param msg: user input data from the configuration mode screen.
+    :return: the setup action on how to continue
+    """
+    global _setup_step
+    global _cfg_add_device
+
+    action = msg.input_values["action"]
+
+    match action:
+        case "add":
+            _cfg_add_device = True
+        case "remove":
+            choice = msg.input_values["choice"]
+            if not config.devices.remove(choice):
+                _LOG.warning("Could not remove device from configuration: %s", choice)
+                return SetupError(error_type=IntegrationSetupError.OTHER)
+            config.devices.store()
+            return SetupComplete()
+        case "reset":
+            config.devices.clear()  # triggers device instance removal
+        case _:
+            _LOG.error("Invalid configuration action: %s", action)
+            return SetupError(error_type=IntegrationSetupError.OTHER)
+
+    _setup_step = SetupSteps.DISCOVER
+    return _user_input_discovery
+
+
+async def _handle_discovery(msg: UserDataResponse) -> RequestUserInput | SetupError:
     """
     Process user data response from the first setup process screen.
 
@@ -142,6 +266,7 @@ async def handle_configuration_mode(msg: UserDataResponse) -> RequestUserInput |
     """
     global _pairing_apple_tv
     global _setup_step
+    global _cfg_add_device
     global _manual_address
     global _discovered_atvs
 
@@ -149,8 +274,6 @@ async def handle_configuration_mode(msg: UserDataResponse) -> RequestUserInput |
     if _pairing_apple_tv:
         await _pairing_apple_tv.disconnect()
         _pairing_apple_tv = None
-    # TODO #11 allow multiple devices!
-    config.devices.clear()  # triggers device instance removal
 
     search_hosts: list[str] | None = None
     dropdown_items = []
@@ -174,6 +297,10 @@ async def handle_configuration_mode(msg: UserDataResponse) -> RequestUserInput |
             device.name,
             device.address,
         )
+        # if we are adding a new device: make sure it's not already configured
+        if _cfg_add_device and config.devices.contains(device.identifier):
+            _LOG.info("Skipping found device %s: already configured", device.identifier)
+            continue
 
         label = f"{device.name} ({device.address})"
         dropdown_items.append({"id": device.identifier, "label": {"en": label}})
@@ -200,7 +327,7 @@ async def handle_configuration_mode(msg: UserDataResponse) -> RequestUserInput |
     )
 
 
-async def handle_device_choice(msg: UserDataResponse) -> RequestUserInput | SetupError:
+async def _handle_device_choice(msg: UserDataResponse) -> RequestUserInput | SetupError:
     """
     Process user data device choice response in a setup process.
 
@@ -214,7 +341,7 @@ async def handle_device_choice(msg: UserDataResponse) -> RequestUserInput | Setu
 
     choice = msg.input_values["choice"]
 
-    atv = discovered_atv_from_identifier(choice)
+    atv = _discovered_atv_from_identifier(choice)
     if atv is None:
         _LOG.error("Chosen Apple TV not found in discovery: %s", choice)
         return SetupError(error_type=IntegrationSetupError.NOT_FOUND)
@@ -245,6 +372,7 @@ async def handle_device_choice(msg: UserDataResponse) -> RequestUserInput | Setu
     if res == 0:
         _LOG.debug("Device provides AirPlay-Code")
         _setup_step = SetupSteps.PAIRING_AIRPLAY
+        # TODO #12 externalize language texts
         return RequestUserInput(
             {
                 "en": "Please enter the shown AirPlay-Code on your Apple TV",
@@ -264,7 +392,7 @@ async def handle_device_choice(msg: UserDataResponse) -> RequestUserInput | Setu
     return RequestUserConfirmation("Please enter the following PIN on your Apple TV: " + res)
 
 
-async def handle_user_data_airplay_pin(msg: UserDataResponse) -> RequestUserInput | SetupError:
+async def _handle_user_data_airplay_pin(msg: UserDataResponse) -> RequestUserInput | SetupError:
     """
     Process user data airplay pairing pin response in a setup process.
 
@@ -299,6 +427,7 @@ async def handle_user_data_airplay_pin(msg: UserDataResponse) -> RequestUserInpu
     if res == 0:
         _LOG.debug("Device provides PIN")
         _setup_step = SetupSteps.PAIRING_COMPANION
+        # TODO #12 externalize language texts
         return RequestUserInput(
             {
                 "en": "Please enter the shown PIN on your Apple TV",
@@ -318,7 +447,7 @@ async def handle_user_data_airplay_pin(msg: UserDataResponse) -> RequestUserInpu
     return RequestUserConfirmation("Please enter the following PIN on your Apple TV: " + res)
 
 
-async def handle_user_data_companion_pin(msg: UserDataResponse) -> SetupComplete | SetupError:
+async def _handle_user_data_companion_pin(msg: UserDataResponse) -> SetupComplete | SetupError:
     """
     Process user data companion pairing pin response in a setup process.
 
@@ -353,8 +482,7 @@ async def handle_user_data_companion_pin(msg: UserDataResponse) -> SetupComplete
         _pairing_apple_tv.get_credentials(),
         _pairing_apple_tv.address,
     )
-    config.devices.add(device)  # triggers ATV instance creation
-    config.devices.store()
+    config.devices.add_or_update(device)  # triggers ATV instance creation
 
     # ATV device connection will be triggered with subscribe_entities request
 
@@ -366,7 +494,7 @@ async def handle_user_data_companion_pin(msg: UserDataResponse) -> SetupComplete
     return SetupComplete()
 
 
-def discovered_atv_from_identifier(identifier: str) -> pyatv.interface.BaseConfig | None:
+def _discovered_atv_from_identifier(identifier: str) -> pyatv.interface.BaseConfig | None:
     """
     Get discovery information from identifier.